# Fix flaky on Ruby 3.2 + Windows and speed up some specs 